+++ ui_playground_generator/lib/src/writor/input/input_field_writor.dart
@@ -12,6 +12,8 @@ class InputFieldWritor {
     InputType.dateTime: 'UiPlaygroundDateTimeInput',
     InputType.edgeInsets: 'UiPlaygroundEdgeInsetsInput',
     InputType.edgeInsetsGeometry: 'UiPlaygroundEdgeInsetsGeometryInput',
+    InputType.borderRadiusGeometry: 'UiPlaygroundBorderRadiusGeometryInput',
+    InputType.borderRadius: 'UiPlaygroundBorderRadiusInput',
     InputType.impaktfullUiAsset: 'UiPlaygroundImpaktfullUiAssetInput',
     InputType.widget: 'UiPlaygroundWidgetInput',
   };