+++ ui_playground/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.5
+
+## Added
+
+- New input types: `UiPlaygroundBorderRadiusInput`, `UiPlaygroundBorderRadiusGeometryInput`
+
 # 0.0.3 - 0.0.4
 
 ## Changed